refactor(build): isolate build functions
BREAKING CHANGE: Simply moving functions from `index.ts` to `build.ts`.

fix #104
