+++ packages/whook-example/bin/build.js
@@ -1,5 +1,5 @@
 #! /usr/bin/env node
 
-const { runBuild } = require('../dist/index');
+const { runBuild } = require('../dist/build');
 
 runBuild();
