+++ packages/whook-example/src/index.ts
@@ -1,15 +1,13 @@
-import Knifecycle, { constant, alsoInject } from 'knifecycle';
+import Knifecycle, { constant } from 'knifecycle';
 import {
   runServer as runBaseServer,
   prepareServer as prepareBaseServer,
   prepareEnvironment as prepareBaseEnvironment,
   initAutoload,
   initAPIDefinitions,
-  initBuildConstants,
 } from '@whook/whook';
 import initHTTPRouter from '@whook/http-router';
 import wrapHTTPRouterWithSwaggerUI from '@whook/swagger-ui';
-import YError from 'yerror';
 import type { DependencyDeclaration, Dependencies } from 'knifecycle';
 
 // Per convention a Whook server main file must export
@@ -82,52 +80,6 @@ export async function prepareEnvironment<T extends Knifecycle<Dependencies>>(
   $.register(
     constant('WHOOK_PLUGINS', ['@whook/cli', '@whook/whook', '@whook/cors']),
   );
-
-  return $;
-}
-
-// Additionnally, 2 others functions can be exported
-//  for building purpose
-
-// The `runBuild` function is intended to build the
-// project
-export async function runBuild(
-  innerPrepareEnvironment = prepareBuildEnvironment,
-): Promise<void> {
-  throw new YError('E_NO_BUILD_IMPLEMENTED');
-
-  // Usually, here you call the installed build
-  // return runBaseBuild(innerPrepareEnvironment);
-}
-
-// The `prepareBuildEnvironment` create the build
-//  environment
-export async function prepareBuildEnvironment<
-  T extends Knifecycle<Dependencies>
->($: T = new Knifecycle() as T): Promise<T> {
-  $ = await prepareEnvironment($);
-
-  // Usually, here you call the installed build env
-  // $ = await prepareBaseBuildEnvironment($);
-
-  // The build often need to know were initializer
-  //  can be found to create a static build and
-  //  remove the need to create an injector
-  $.register(
-    constant('INITIALIZER_PATH_MAP', {
-      ENV: '@whook/whook/dist/services/ProxyedENV',
-      apm: '@whook/http-transaction/dist/services/apm',
-      obfuscator: '@whook/http-transaction/dist/services/obfuscator',
-      errorHandler: '@whook/http-router/dist/services/errorHandler',
-      log: 'common-services/dist/log',
-      time: 'common-services/dist/time',
-      delay: 'common-services/dist/delay',
-    }),
-  );
-
-  // Finally, some constants can be serialized instead of being
-  //  initialized in the target build saving some time at boot
-  $.register(alsoInject(['API_DEFINITIONS'], initBuildConstants));
 
   return $;
 }
